backend/requirements.txt
langchain
langchain-community
langchain-core
langsmith
openai
ipykernel
pydantic
pydantic_core
langchain_openai
arxiv
pymupdf
bs4
lxml
wikipedia
pypdf
unstructured
tiktoken
langchain-groq
transformers
pdfminer.six
pypdf
langchain-google-genai
langchain-mistralai
langchain-chroma
faiss-cpu
fastapi
python-multipart
uvicorn
python-dotenv
langchain-pymupdf4llm
pillow